Delay loading of most shape settings data
