+++ client/package.json
@@ -62,6 +62,7 @@
         "sass": "^1.58.1",
         "typescript": "^4.9.5",
         "upath": "^2.0.1",
+        "v-lazy-show": "^0.2.0",
         "vite": "^4.1.1",
         "vitest": "^0.28.5",
         "vue-eslint-parser": "^9.1.0",
