+++ client/vite.config.ts
@@ -4,11 +4,12 @@ import path from "path";
 import { defineConfig } from "vite";
 import vue from "@vitejs/plugin-vue";
 import vueI18n from "@intlify/unplugin-vue-i18n/vite";
+import { transformLazyShow } from "v-lazy-show";
 
 // https://vitejs.dev/config/
 export default defineConfig({
     plugins: [
-        vue(),
+        vue({ template: { compilerOptions: { nodeTransforms: [transformLazyShow] } } }),
         vueI18n({
             include: path.resolve(__dirname, "./src/locales/**"),
         }),
